clean up block title
This reverts commit 7e9d6462ff5d1908abacabfac320f5bef5178404.
