+++ addon/components/nypr-m-block/title-layout.js
@@ -1,0 +1,8 @@
+import Component from '@ember/component';
+import layout from '../../templates/components/nypr-m-block/title-layout';
+
+
+export default Component.extend({
+  layout,
+  tagName: '',
+});
