+++ addon/templates/components/nypr-m-block/title.hbs
@@ -1,23 +1,27 @@
 {{#if (eq @h2 '')}}
-  {{#nypr-m-block/h2
-    route=@route
-    url=@url
-    hasGallery=@hasGallery
-    hasVideo=@hasVideo
-    hasAudio=@hasAudio
-    data-test-block-title=true
-  }}
-    {{yield}}
-  {{/nypr-m-block/h2}}
+  <h2 class="c-block__title {{this.iconClassNames}}">
+    {{#nypr-m-block/title-layout
+      route=@route
+      url=@url
+      hasGallery=@hasGallery
+      hasVideo=@hasVideo
+      hasAudio=@hasAudio
+      data-test-block-title=true
+    }}
+      {{yield}}
+    {{/nypr-m-block/title-layout}}
+  </h2>
 {{else if (eq @h3 '')}}
-  {{#nypr-m-block/h3
-    route=@route
-    url=@url
-    hasGallery=@hasGallery
-    hasVideo=@hasVideo
-    hasAudio=@hasAudio
-    data-test-block-title=true
-  }}
-    {{yield}}
-  {{/nypr-m-block/h3}}
+  <h3 class="c-block__title {{this.iconClassNames}}">
+    {{#nypr-m-block/title-layout
+      route=@route
+      url=@url
+      hasGallery=@hasGallery
+      hasVideo=@hasVideo
+      hasAudio=@hasAudio
+      data-test-block-title=true
+    }}
+      {{yield}}
+    {{/nypr-m-block/title-layout}}
+  </h3>
 {{/if}}
