+++ app/components/nypr-m-block/title-layout.js
@@ -1,0 +1,1 @@
+export { default } from 'nypr-design-system/components/nypr-m-block/title-layout';
